IT1566/Assignment1/requirements.txt
flask
flask-login
flask-sqlalchemy
jinja2
python-dotenv